Fix crash + textures of Sri Lankan Junglefowl
In the initial commit, forgot to include the textures and to fully replace the references to Grey Junglefowl in registration. Additionally updated the Taggable API class to include a type-specific getTag() method, so that an integer/float/etc. can be directly queried and avoid causing an NPE upon trying to cast "null" to "Integer" or "Float" or etc.