Possible fix for ASM NPE spam of TooltipEvent in main menu
I've noticed a few times while testing that the log will be spammed by a NullPointerException while firing Forge's TooltipEvent after I exit a world and return to the main menu. The game does not crash, but it constantly spams the log with the locations of all TooltipEvent handlers. Might not even be my fault, but I'm including a new nullcheck. Better safe than spammy.